Create a useConversations file to use as a hook
